requirements.txt
python-telegram-bot==22.7
google-genai==1.68.0
python-dotenv==1.2.2
